Renamed SHA3.{h,cpp} files.
