+++ test/libsolidity/AnalysisFramework.cpp
@@ -29,7 +29,7 @@
 
 #include <libsolidity/parsing/Scanner.h>
 
-#include <libdevcore/SHA3.h>
+#include <libdevcore/Keccak256.h>
 
 #include <boost/test/unit_test.hpp>
