+++ test/libsolidity/SolidityTypes.cpp
@@ -22,7 +22,7 @@
 
 #include <libsolidity/ast/Types.h>
 #include <libsolidity/ast/AST.h>
-#include <libdevcore/SHA3.h>
+#include <libdevcore/Keccak256.h>
 #include <boost/test/unit_test.hpp>
 
 using namespace std;
